Generic settings resource revamped
